Allow setting image protocol in config file
